requirements.txt
google-generativeai>=0.3.0
streamlit>=1.22.0
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0
uuid>=1.30
python-dotenv>=0.19.0
